core/requirements/base.txt
asgiref==3.9.1
attrs==25.3.0
black==25.1.0
click==8.2.1
Django==5.2.6
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
ruff==0.12.12
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
